Move MQTT and InfluxDB code to separate header files
